fuse: fake getattr so we don't fetch file on each stat
Utilities like cp(1) that compare inode numbers currently fail
to work because they detect the file being replaced while being copied.

We can workaround this by setting use_ino to zero, but we are left with
the problem that stat(2), which is supposed to be a cheap operation,
downloads the file fully each time. TFTP doesn't provide means to get
file size and we don't want to cache files as they can change on the
server. Thus just return some fake contents, which should be acceptable
to most programs.

This hardcodes the maximum file size as 2G, but you shouldn't download
more than that over TFTP anyway...