tftp: use tsize extension to support stat(2)
While faking the file size works for cp(1), dnsmasq doesn't like the
short read on files < 2G and aborts transfer when reading from a mounted
tftpfs. Apparently there is an extension to query size in advance, so
make use of it.

For widest possible compatibility, a fall back to downloading the file
could be re-added, but for now just require tsize support.